Remove draw_map fucntion
The draw_map function is modified and moved to the Game class.
Some new function called delay is used for showing information slower than the player will feel comfortable while playing the game. Because the game in the console many pieces of information are printed too fast it will be a mess and the player can't catch them correctly. Another reason is the time() is used as a seed for the random result,the deday will help that.