refactor: add new state named 'player_type'. For reducing duplicate code.
Now, players are classified into two types: auto and manual. The game
process applies for both types the same. The only thing is the "auto"
player will automatically fill answer and the manual player must enter
it. So we can reuse the code of the processing part instead of having to
write code for both objects.