bug: importerror and int cannot iterator
importerror caused by race_view.py and race.py. Removing the 'main.' part will escaping
from recursion import

Regarding the error: "int object is not iterable". This is due to the
data type of the Race property. The property named boxes_positions
contains the data in a tuple of integers. In case the tuple has only one
integer, it is considered to be of the class integer. To solve the
problem I change to List type.